Rename analog_state_controller->analog_controller
