+++ analog_controller/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>analog_state_controller</name>
+  <name>analog_controller</name>
   <version>0.0.0</version>
   <description>Controller to publish the state of analog sensors</description>
 
@@ -21,6 +21,6 @@
   <depend>roscpp</depend>
 
   <export>
-    <controller_interface plugin="${prefix}/analog_state_controller_plugins.xml"/>
+    <controller_interface plugin="${prefix}/analog_controller_plugins.xml"/>
   </export>
 </package>
