+++ frc_control/package.xml
@@ -18,7 +18,7 @@
   <exec_depend>frc_msgs</exec_depend>
   <exec_depend>frc_robot_hw</exec_depend>
   <exec_depend>frc_robot_sim</exec_depend>
-  <exec_depend>analog_state_controller</exec_depend>
+  <exec_depend>analog_controller</exec_depend>
   <exec_depend>binary_state_controller</exec_depend>
   <exec_depend>ternary_state_controller</exec_depend>
   <exec_depend>pdp_state_controller</exec_depend>
