+++ frc_robot_hw/CMakeLists.txt
@@ -12,7 +12,7 @@ find_package(catkin REQUIRED COMPONENTS
   transmission_interface
   urdf
 
-  analog_state_controller
+  analog_controller
   binary_state_controller
   frc_msgs
   ternary_state_controller
@@ -33,7 +33,7 @@ catkin_package(
     transmission_interface
     urdf
 
-    analog_state_controller
+    analog_controller
     binary_state_controller
     frc_msgs
     ternary_state_controller
